% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_files}
\alias{add_files}
\title{Add files to a gist object}
\usage{
add_files(.gist, files)
}
\description{
Add files to a gist object
}
\examples{
\dontrun{
add_files(files="~/stuff.Rmd")
add_files(files=c("~/bunnycheese.txt","~/texas.txt","~/stuff.Rmd"))
}
}

