% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Open a gist on GitHub}
\usage{
browse(gist, what = "html")
}
\arguments{
\item{gist}{A gist object or something that can be coerced to a gist object.}

\item{what}{One of html (default), json, forks, commits, or comments.}
}
\description{
Open a gist on GitHub
}

