% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delete}
\alias{delete}
\title{Delete a gist}
\usage{
delete(gist, ...)
}
\arguments{
\item{...}{Curl options passed on to httr::GET}

\item{id}{Gist id}

\item{verbose}{Print informative messages (default TRUE)}
}
\description{
Delete a gist
}
\examples{
\dontrun{
gists("minepublic")[[1]] \%>\% delete()
}
}

