% Generated by roxygen2 (4.0.2): do not edit by hand
\name{embed}
\alias{embed}
\title{Get embed script for a gist}
\usage{
embed(gist)
}
\arguments{
\item{gist}{A gist object or something that can be coerced to a gist object.}
}
\description{
Get embed script for a gist
}
\examples{
\donttest{
gists()[[1]] \%>\% embed()
}
}

