% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forks.R
\name{fork}
\alias{fork}
\title{Fork a gist}
\usage{
fork(gist, ...)
}
\arguments{
\item{gist}{A gist object or something coerceable to a gist}

\item{...}{Further named args to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
A gist class object
}
\description{
Fork a gist
}
\examples{
\dontrun{
# fork a gist
w <- gists()[[1]] \%>\% fork()

# browse to newly forked gist
browse(w)
}
}
