% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forks.R
\name{forks}
\alias{forks}
\title{List forks on a gist}
\usage{
forks(gist, page = NULL, per_page = 30, ...)
}
\arguments{
\item{gist}{A gist object or something coerceable to a gist}

\item{page}{(integer) Page number to return.}

\item{per_page}{(integer) Number of items to return per page. Default 30.
Max 100.}

\item{...}{Further named args to \code{\link[httr:GET]{httr::GET()}}}
}
\value{
A list of gist class objects
}
\description{
List forks on a gist
}
\examples{
\dontrun{
gist(id='1642874') %>% forks(per_page=2)
gist(id = "8172796") %>% forks()

# pass in a url
gist("https://gist.github.com/expersso/4ac33b9c00751fddc7f8") %>% forks 
}
}
