% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gist_commits}
\alias{gist_commits}
\title{List gist commits}
\usage{
gist_commits(.gist, id = NULL, page = NULL, per_page = 30, ...)
}
\arguments{
\item{.gist}{gist object}

\item{id}{Gist id.}

\item{page}{(integer) Page number to return.}

\item{per_page}{(integer) Number of items to return per page. Default 30. Max 100.}

\item{...}{Further named args to \code{\link[httr]{GET}}}
}
\description{
List gist commits
}
\examples{
\dontrun{
gist_commits(id='cf5d2e572faafb4c6d5f')
# just 1 commit
gist_commits(id='cf5d2e572faafb4c6d5f', per_page=1)
}
}

