% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gist_create}
\alias{gist_create}
\title{Create a gist}
\usage{
gist_create(files, description = "", public = TRUE, browse = TRUE, ...)
}
\arguments{
\item{files}{Files to upload}

\item{description}{(character) Brief description of gist (optional)}

\item{public}{(logical) Whether gist is public (default: TRUE)}
}
\description{
Create a gist
}
\examples{
\donttest{
gist_create(files="~/stuff.md", description='a new cool gist')
gist_create(files=c("~/spocc_sp.Rmd","~/spocc_sp.md"), description='spocc demo files')
}
}

