% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gist_create_git.R
\name{gist_create_git}
\alias{gist_create_git}
\title{Create a gist via git instead of the HTTP API}
\usage{
gist_create_git(files = NULL, description = "", public = TRUE,
  browse = TRUE, knit = FALSE, code = NULL, filename = "code.R",
  knitopts = list(), renderopts = list(), include_source = FALSE,
  artifacts = FALSE, imgur_inject = FALSE, ...)
}
\arguments{
\item{files}{Files to upload}

\item{description}{(character) Brief description of gist (optional)}

\item{public}{(logical) Whether gist is public (default: TRUE)}

\item{browse}{(logical) To open newly create gist in default browser (default: TRUE)}

\item{knit}{(logical) Knit code before posting as a gist? If the file has a \code{.Rmd}
or \code{.Rnw} extension, we run the file with \code{link[knitr]{knit}}, and if it has
a \code{.R} extension, then we use \code{\link[rmarkdown]{render}}}

\item{code}{Pass in any set of code. This can be a single R object, or many lines of code
wrapped in quotes, then curly brackets (see examples below).}

\item{filename}{Name of the file to create, only used if \code{code} parameter is used. Default
to \code{code.R}}

\item{knitopts,renderopts}{(list) List of variables passed on to \code{link[knitr]{knit}},
or \code{\link[rmarkdown]{render}}}

\item{include_source}{(logical) Only applies if \code{knit=TRUE}. Include source file in the
gist in addition to the knitted output.}

\item{artifacts}{(logical) Include artifacts or not, Ignored for now. Default: FALSE}

\item{imgur_inject}{(logical) Inject \code{\link[knitr]{imgur_upload}} into your
\code{.Rmd} file to upload files to \url{http://imgur.com/}. This will be ignored
if the file is a sweave/latex file because the rendered pdf can't be uploaded
anyway. Default: FALSE}

\item{...}{Further args passed on to \code{link[httr]{POST}}}
}
\description{
Create a gist via git instead of the HTTP API
}
\examples{
\dontrun{
gist_create_git(files = "~/gitgist/stuff.md")
gist_create_git(files = c("~/gitgist/stuff.md", "~/gitgist/icanhazallthedata.md"))
gist_create_git(files = "~/gitgist")

# binary files
gist_create_git(files = "~/gitgist/ropensci_add.png")

# knit files first, then push up
gist_create_git(files = "~/gitgistknit/plots.Rmd", knit = TRUE)

# collect all/any artifacts from knitting process
gist_create_git("~/gitgistartifacts/artifacts_eg1.Rmd", knit = TRUE,
   artifacts = TRUE)

# from code
gist_create_git(code={'
x <- letters
numbers <- runif(8)
numbers

[1] 0.3229318 0.5933054 0.7778408 0.3898947 0.1309717 0.7501378 0.3206379 0.3379005
'}, filename="my_cool_code.R")
}
}

