% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gist_delete}
\alias{gist_delete}
\title{Delete a gist}
\usage{
gist_delete(id, verbose = TRUE)
}
\arguments{
\item{id}{Gist id}

\item{verbose}{Print messages (default TRUE)}
}
\description{
Delete a gist
}
\examples{
\dontrun{
gist_delete(id='42095d9a661b0eefc694')
}
}

