% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gist_get}
\alias{gist_get}
\title{List gists}
\usage{
gist_get(id = NULL, what = "public", since = NULL, page = NULL,
  per_page = 30, verbose = TRUE, callopts = list())
}
\arguments{
\item{id}{Gist id.}

\item{what}{(character) What gists to return. One of public, minepublic, mineall, or starred.
If an id is given for a gist, this parameter is ignored.}

\item{since}{(character) A timestamp in ISO 8601 format: YYYY-MM-DDTHH:MM:SSZ. Only gists
updated at or after this time are returned.}

\item{page}{(integer) Page number to return.}

\item{per_page}{(integer) Number of items to return per page. Default 30. Max 100.}

\item{verbose}{Print informative messages (default TRUE)}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
You can list public gists, your public gists, or all your gists
}
\examples{
\dontrun{
# Public gists
gist_get()
gist_get(per_page=1)
gist_get(page=3)
# Public gists created since X time
gist_get(since='2014-05-26T00:00:00Z')
# Your public gists
gist_get(what='minepublic')
gist_get(what='minepublic', per_page=2)
# Your private and public gists
gist_get(what='mineall')
# A single gist
gist_get(id='cf5d2e572faafb4c6d5f')
# Your starred gists
gist_get(what='starred')
# pass in curl options
gist_get(per_page=1, callopts=verbose())
gist_get(per_page=1, callopts=timeout(seconds = 0.5))
}
}

