% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gist_oauth}
\alias{gist_oauth}
\title{Authorize with GitHub.}
\usage{
gist_oauth(app = gistr_app, reauth = FALSE)
}
\arguments{
\item{app}{An \code{\link[httr]{oauth_app}} for DO. The default uses the
standard ROpenSci application.}

\item{reauth}{(logical) Force re-authorization?}
}
\description{
This function is run automatically to allow analogsea to access your
GitHub account.
}
\details{
There are two ways to authorise gistr to work with your GitHub account:
\itemize{
\item Generate a personal access token at
  \url{https://cloud.digitalocean.com/settings/tokens/new} and
  record in the \code{DO_PAT} envar.

\item Interactively login into your DO account and authorise with
  OAuth.
}

Using \code{DO_PAT} is recommended.
}
\examples{
\donttest{
gist_oauth()
}
}

