% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gist_save.R
\name{gist_save}
\alias{gist_save}
\title{Save a gist to disk}
\usage{
gist_save(gist, path = ".", open_files = FALSE)
}
\arguments{
\item{gist}{A gist object or something coerceable to a gist}

\item{path}{Path to write to, a directory, not a file b/c a gist can contain
many files. File names will be the same as given in the gist}

\item{open_files}{(logical) If \code{TRUE}, open files in your editor/R GUI.
Default: \code{FALSE}}
}
\value{
file(s) path
}
\description{
Save a gist to disk
}
\examples{
\dontrun{
gist("a65ac7e56b7b3f746913") \%>\% gist_save()
gist("https://gist.github.com/expersso/4ac33b9c00751fddc7f8") \%>\% gist_save()
}
}

