% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gist_star}
\alias{gist_star}
\title{Star a gist}
\usage{
gist_star(id, what = "star", verbose = TRUE)
}
\arguments{
\item{id}{Gist id}

\item{what}{(character) One of star (default), unstar, or check.}

\item{verbose}{Print informative messages (default TRUE)}
}
\description{
Star a gist
}
\examples{
\dontrun{
gist_star(id='7698648')
gist_star(id='7698648', what='check')
gist_star(id='7698648', what='unstar')
gist_star(id='7698648', what='check')
}
}

