% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gist_update}
\alias{gist_update}
\title{Update/modify a gist}
\usage{
gist_update(gist, id, description = "", public = TRUE)
}
\arguments{
\item{id}{Gist id}

\item{verbose}{Print messages (default TRUE)}
}
\description{
Update/modify a gist
}
\examples{
\dontrun{
gist_update(id='42095d9a661b0eefc694')
}
}

