% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gistr_map.R
\name{gistr_map}
\alias{gistr_map}
\title{Opens a full screen map after uploading a geojson file}
\usage{
gistr_map(file = NULL, gist_object, browse = TRUE)
}
\arguments{
\item{file}{Must be a geojson file}

\item{gist_object}{An object generated by \code{gist_id}}

\item{browse}{Default is \code{TRUE}. Set to \code{FALSE} if you don't want to automatically browse to the URL.}
}
\description{
Takes a gist object and a input geojson file name and renders fullscreen map
}
\examples{
\dontrun{
library(ecoengine)
x <- ee_observations(genus = "vulpes")
ee_map(x, dest = "~/Desktop")
gist_id <- gist_create(files = "~/Desktop/temp.geojson", browse = FALSE)
gistr_map(file = "temp.geojson", gist_id)

}
}

