% Generated by roxygen2 (4.0.2): do not edit by hand
\name{post}
\alias{post}
\title{Post a gist object}
\usage{
post(gist, description = "", public = TRUE, verbose = TRUE,
  browse = TRUE, ...)
}
\description{
Post a gist object
}
\examples{
\dontrun{
post(description='My gist!', public=TRUE)
}
}

