% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rate_limit}
\alias{rate_limit}
\title{Get rate limit information}
\usage{
rate_limit(.gist, ...)
}
\arguments{
\item{.gist}{gist object}

\item{...}{Further named args to \code{\link[httr]{GET}}}
}
\description{
Get rate limit information
}
\examples{
\dontrun{
gist() \%>\% rate_limit
gist() \%>\% rate_limit(config=verbose())
}
}

