% Generated by roxygen2 (4.0.2): do not edit by hand
\name{star}
\alias{star}
\alias{star_check}
\alias{unstar}
\title{Star a gist}
\usage{
star(.gist, id, ...)

unstar(.gist, id, ...)

star_check(.gist, id, ...)
}
\arguments{
\item{.gist}{gist object}

\item{id}{Gist id}

\item{...}{Curl options passed on to httr::GET}

\item{verbose}{Print informative messages (default TRUE)}
}
\description{
Star a gist
}
\examples{
\dontrun{
id <- '7ddb9810fc99c84c65ec'
gist() \%>\% star(id = id)
gist() \%>\% star_check(id = id)
gist() \%>\% unstar(id = id)
gist() \%>\% star_check(id = id)
gist() \%>\%
  star(id = id) \%>\%
  star_check(id = id)
}
}

