% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tbl.R
\name{tabl}
\alias{tabl}
\alias{tabl.gist}
\alias{tabl.list}
\title{Make a table from gist class or list of gist class objects}
\usage{
tabl(x)

\method{tabl}{gist}(x)

\method{tabl}{list}(x)
}
\arguments{
\item{x}{Either a gist class object or a list of many gist class objects}
}
\description{
Make a table from gist class or list of gist class objects
}
\examples{
\dontrun{
x <- as.gist('f1403260eb92f5dfa7e1')
tabl(x)

# from a list
ss <- gists('minepublic')
tabl(ss[1:3])
## manipulate with dplyr
library("dplyr")
tabl(gists("minepublic")[1:10]) \%>\%
  select(id, description, owner_login) \%>\%
  filter(grepl("gist gist gist", description)) \%>\%
  delete
}
}

