% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tbl.R
\name{tbl}
\alias{tbl}
\alias{tbl.gist}
\alias{tbl.list}
\title{Make a table from gist class or list of gist class objects}
\usage{
tbl(x)

\method{tbl}{gist}(x)

\method{tbl}{list}(x)
}
\arguments{
\item{x}{Either a gist class object or a list of many gist class objects}
}
\description{
Make a table from gist class or list of gist class objects
}
\examples{
\dontrun{
x <- as.gist('f1403260eb92f5dfa7e1')
tbl(x)

# from a list
ss <- gists('minepublic')
tbl(ss[1:3])
## manipulate with dplyr
library("dplyr")
tbl(gists("minepublic")[1:4]) \%>\%
  select(id, description, owner_login) \%>\%
  filter(grepl("gist gist gist", description)) \%>\%
  delete
}
}

