% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\alias{delete.gist}
\alias{delete.tbl}
\title{Delete a gist}
\usage{
delete(x, ...)

\method{delete}{gist}(x, ...)

\method{delete}{tbl}(x, ...)
}
\arguments{
\item{x}{A gist object or something coerceable to a gist}

\item{...}{Curl options passed on to httr::GET}
}
\description{
Delete a gist
}
\examples{
\dontrun{
gists("minepublic")[[29]] \%>\% delete()
}
}

