% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delete.R
\name{delete}
\alias{delete}
\title{Delete a gist}
\usage{
delete(gist, ...)
}
\arguments{
\item{gist}{A gist object or something coerceable to a gist}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Delete a gist
}
\examples{
\dontrun{
gists("minepublic")[[29]] \%>\% delete()
}
}

