% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embed.R
\name{embed}
\alias{embed}
\title{Get embed script for a gist}
\usage{
embed(gist)
}
\arguments{
\item{gist}{A gist object or something that can be coerced to a gist object.}
}
\description{
Get embed script for a gist
}
\examples{
\dontrun{
gists()[[1]] \%>\% embed()
}
}

