% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_files}
\alias{add_files}
\alias{delete_files}
\alias{edit_files}
\alias{rename_files}
\title{Add files to a gist object}
\usage{
add_files(gist, ...)

edit_files(gist, ...)

delete_files(gist, ...)

rename_files(gist, ...)
}
\arguments{
\item{gist}{A gist object or something coerceable to a gist}

\item{...}{Curl options passed on to httr::GET}
}
\description{
Add files to a gist object
}
\examples{
\donttest{
add_files("~/stuff.Rmd")
edit_files()
delete_files()
rename_files()
}
}

