% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forks.R
\name{fork}
\alias{fork}
\title{Fork a gist}
\usage{
fork(gist, ...)
}
\arguments{
\item{gist}{A gist object or something coerceable to a gist}

\item{...}{Further named args to \code{\link[httr]{GET}}
#'}
}
\value{
A gist class object
}
\description{
Fork a gist
}
\examples{
\dontrun{
# fork a gist
gists()[[1]] \%>\% fork()

# browse to newly forked gist
gist(id='0831f3fbd83ac4d46451') \%>\% fork() \%>\% browse()

# extract the last one
gist(id='1642874') \%>\%
 forks() \%>\%
 extract(length(.))
}
}

