% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gist.R
\name{gist}
\alias{as.gist}
\alias{gist}
\title{Get a gist}
\usage{
gist(id, ...)

as.gist(x)
}
\arguments{
\item{id}{(character) A gist id}

\item{...}{Curl options passed on to httr::GET}

\item{x}{Object to coerce. Can be an integer (gist id), string
(gist id), a gist, or an list that can be coerced to a gist.}
}
\description{
Get a gist
}
\examples{
\donttest{
gist(id = 'f1403260eb92f5dfa7e1')

as.gist('f1403260eb92f5dfa7e1')
as.gist(10)
as.gist(gist('f1403260eb92f5dfa7e1'))

ids <- sapply(gists(), "[[", "id")
gist(ids[1])
gist(ids[2])
gist(ids[3])
gist(ids[4])

gist(ids[1]) \%>\% browse()
}
}

