% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gist}
\alias{gist}
\title{Get a gist}
\usage{
gist(id, ...)
}
\arguments{
\item{id}{(character) A gist id}

\item{...}{Curl options passed on to httr::GET}
}
\description{
Get a gist
}
\examples{
\donttest{
gist(id = 'f1403260eb92f5dfa7e1')

as.gist('f1403260eb92f5dfa7e1')
as.gist(10)
as.gist(gist('f1403260eb92f5dfa7e1'))

ids <- sapply(gists(), "[[", "id")
gist(ids[1])
gist(ids[2])
gist(ids[3])
gist(ids[4])

gist(ids[1]) \%>\% browse()
}
}

