% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gist_create}
\alias{gist_create}
\title{Post a file as a Github gist}
\usage{
gist_create(gist, description = "", public = TRUE, verbose = TRUE)
}
\arguments{
\item{gist}{(character) Path to one or more files.}

\item{description}{(character) Brief description of gist (optional)}

\item{public}{(logical) Whether gist is public (default: TRUE)}
}
\value{
Posts your file as a gist on your account, and prints out the url for the
gist and for embedding the gist in the console.
}
\description{
You will be asked ot enter you Github credentials (username, password) during
each session, but only once for each session. Alternatively, you could enter
your credentials into your .Rprofile file with the entries

\enumerate{
 \item options(github.username = 'your_github_username')
 \item options(github.password = 'your_github_password')
}

then \code{gist} will simply read those options.
}
\examples{
\dontrun{
gist_create(files="stuff.md", description='My gist!', public=TRUE)
}
}

