% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gist_create}
\alias{gist_create}
\title{Post a file as a Github gist}
\usage{
gist_create(gist, description = "", public = TRUE)
}
\arguments{
\item{gist}{An object}

\item{description}{brief description of gist (optional)}

\item{public}{whether gist is public (default: TRUE)}
}
\value{
Posts your file as a gist on your account, and prints out the url for the
gist itself in the console.
}
\description{
You will be asked ot enter you Github credentials (username, password) during
each session, but only once for each session. Alternatively, you could enter
your credentials into your .Rprofile file with the entries

\enumerate{
 \item options(github.username = 'your_github_username')
 \item options(github.password = 'your_github_password')
}

then \code{gist} will simply read those options.

\code{gist} was modified from code in the rCharts package by Ramnath Vaidyanathan
}
\examples{
\dontrun{
gist_create(files="stuff.md", description='My gist!', public=TRUE)
}
}

