% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gist_delete}
\alias{gist_delete}
\title{Delete a gist}
\usage{
gist_delete(id, verbose = TRUE, callopts = list())
}
\arguments{
\item{id}{Gist id}

\item{verbose}{Print informative messages (default TRUE)}

\item{...}{Curl options passed on to httr::GET}
}
\description{
Delete a gist
}
\examples{
\dontrun{
gist_delete(id='42095d9a661b0eefc694')
}
}

