% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gist_star}
\alias{gist_star}
\title{Star a gist}
\usage{
gist_star(id, what = "star", verbose = TRUE, callopts = list())
}
\arguments{
\item{id}{Gist id}

\item{what}{(character) One of star (default), unstar, or check.}

\item{verbose}{Print informative messages (default TRUE)}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
Star a gist
}
\examples{
\dontrun{
gist_star(id='7698648')
gist_star(id='7698648', what='check')
gist_star(id='7698648', what='unstar')
gist_star(id='7698648', what='check')
}
}

