% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gist_star}
\alias{gist_star}
\title{Star a gist}
\usage{
gist_star(id, verbose = TRUE)
}
\arguments{
\item{id}{Gist id}

\item{verbose}{Print informative messages (default TRUE)}
}
\description{
Star a gist
}
\examples{
\dontrun{
gist_star(id='2f9eb6b72730fed2f061')
}
}

