% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gist_update}
\alias{gist_update}
\title{Update/modify a gist}
\usage{
gist_update(id, gist = NULL, description = "", public = TRUE,
  verbose = TRUE)
}
\arguments{
\item{id}{Gist id}

\item{gist}{(character) Path to one or more files.}

\item{description}{(character) Brief description of gist (optional)}

\item{public}{(logical) Whether gist is public (default: TRUE)}

\item{verbose}{Print informative messages (default TRUE)}
}
\description{
Update/modify a gist
}
\examples{
\dontrun{
gist_update(id='42095d9a661b0eefc694')
}
}

