% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gistr-package.R
\docType{package}
\name{gistr-package}
\alias{gistr-package}
\alias{gistr}
\title{R client for GitHub gists}
\description{
R client for GitHub gists.
}
\details{
gistr allows you to peform actions on gists, including listing, forking,
starring, creating, deleting, updating, etc.

There are two ways to authorise gistr to work with your GitHub account:
\itemize{
\item Generate a personal access token (PAT) at
\url{https://docs.github.com/en/github/authenticating-to-github/creating-a-personal-access-token}
and record it in the \code{GITHUB_PAT} envar.
\item Interactively login into your GitHub account and authorise with OAuth.
}

Using the \code{GITHUB_PAT} is recommended.
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Ramnath Vaidyanathan \email{ramnath.vaidya@gmail.com}

Karthik Ram \email{karthik.ram@gmail.com}
}
\keyword{package}
