% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gistr-package.R
\docType{package}
\name{gistr-package}
\alias{gistr}
\alias{gistr-package}
\title{R client for GitHub gists}
\description{
R client for GitHub gists.
}
\details{
gistr allows you to peform actions on gists, including listing, forking, starring, creating,
deleting, updating, etc.

There are two ways to authorise gistr to work with your GitHub account:

\itemize{
 \item Generate a personal access token (PAT) at
 \url{https://help.github.com/articles/creating-an-access-token-for-command-line-use}
 and record it in the \code{GITHUB_PAT} envar.
 \item Interactively login into your GitHub account and authorise with OAuth.
}

Using the \code{GITHUB_PAT} is recommended.

We also provide an interface to GitHub Gists using git, bypassing the http API.
This path is much easier if you already use git locally. If not, you'll have to
setup some git options. See \url{https://github.com/ropensci/git2r#configuration}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

Ramnath Vaidyanathan \email{ramnath.vaidya@gmail.com}

Karthik Ram \email{karthik.ram@gmail.com}
}
\keyword{package}

