% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gists}
\alias{gists}
\title{List gists}
\usage{
gists(.gist, id = NULL, what = "public", since = NULL, page = NULL,
  per_page = NULL, verbose = TRUE, ...)
}
\arguments{
\item{id}{Gist id.}

\item{what}{(character) What gists to return. One of public, minepublic, mineall, or starred.
If an id is given for a gist, this parameter is ignored.}

\item{since}{(character) A timestamp in ISO 8601 format: YYYY-MM-DDTHH:MM:SSZ. Only gists
updated at or after this time are returned.}

\item{page}{(integer) Page number to return.}

\item{per_page}{(integer) Number of items to return per page. Default 30. Max 100.}

\item{verbose}{Print informative messages (default TRUE)}

\item{callopts}{Curl options passed on to httr::GET}
}
\description{
You can list public gists, your public gists, or all your gists
}
\examples{
\dontrun{
# Public gists
gists()
gists(per_page=1)
gists(page=3)
# Public gists created since X time
gists(since='2014-05-26T00:00:00Z')
# Your public gists
gists(what='minepublic')
gists(what='minepublic', per_page=2)
# Your private and public gists
gists(what='mineall')
# A single gist
gists(id='cf5d2e572faafb4c6d5f')
# Your starred gists
gists(what='starred')
# pass in curl options
gists(per_page=1, config=verbose())
gists(per_page=1, config=timeout(seconds = 0.5))
}
}

