% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rate_limit}
\alias{rate_limit}
\title{Get rate limit information}
\usage{
rate_limit(...)
}
\arguments{
\item{...}{Named args to \code{\link[httr]{GET}}}
}
\description{
Get rate limit information
}
\examples{
\donttest{
rate_limit()
rate_limit(config=verbose())
}
}

