% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{rate_limit}
\alias{rate_limit}
\title{Get rate limit information}
\usage{
rate_limit(...)
}
\arguments{
\item{...}{Named args to \link[crul:verb-GET]{crul::verb-GET}}
}
\description{
Get rate limit information
}
\examples{
\dontrun{
rate_limit()
}
}
