% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rate_limit.R
\name{rate_limit}
\alias{rate_limit}
\title{Get rate limit information}
\usage{
rate_limit(...)
}
\arguments{
\item{...}{Named args to \code{\link[httr]{GET}}}
}
\description{
Get rate limit information
}
\examples{
\dontrun{
rate_limit()
rate_limit(config=verbose())
}
}

